backend/analytics/requirements.txt
appnope==0.1.0
astroid==2.0.4
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
flake8==3.6.0
idna==2.7
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
joblib==0.13.0
jsonschema==2.6.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==3.0.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.1.6
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.7.0
numpy==1.15.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==2.0.6
ptyprocess==0.6.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.2
python-dateutil==2.7.4
python-dotenv==0.9.1
pytz==2018.6
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xgboost==0.81
